requirements.txt
beautifulsoup4==4.12.2
fastapi==0.104.1
Flask==3.0.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
httpcore==1.0.2
httplib2==0.22.0
imap-tools==1.5.0
importlib-metadata==6.8.0
langchain==0.0.339
numpy==1.24.4
openai==0.28.0
pandas==2.1.3
python-dotenv==1.0.0